flashcards are now added to ALL decks for which they are tagged.
The side effect of this is that cards are double-counted in the parentTag deck if they are tagged with both #parentTag and #parentTag/childTag. This seems odd but if that's how the user is choosing to do it, I think we should respect that. Will look into making a setting to control this behavior next.