css fix for phone screens
a fix for phone screens as I encountered a very broken view when trying to see the flashcards on a phone.

This fixes the displayproblem with ios devices. On android phones this bug doesn't seem to appear. BUT this also changes how much of the display is taken up. Because on PC not the whole vertical space is taken I also made a bit of space above and below.